노드 종류
